src/requirements.txt
fastapi
uvicorn
openai-whisper
pydantic
python-multipart